IPT/requirements.txt
Flask==2.1.2
pandas==1.4.2
numpy==1.22.3
matplotlib==3.5.2
seaborn==0.11.2
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.1
